Dirty hack while I look up how to hook the real structure (or jump directly to PCRE) to avoid buffer overflows. Moved everything into java.util.regex and changed the interface so is compatible with java, changed the few tests I added to work with that, too
