+++ javalib/src/main/scala/java/util/regex/Matcher.scala
@@ -1,4 +1,4 @@
-package scala.scalanative.posix
+package java.util.regex
 
 trait MatchResult {
   def groupCount(): Int
@@ -12,19 +12,19 @@ trait MatchResult {
   def group(group: Int): String
 }
 
-final class RegexMatcher private[posix] (private var _pattern: PosixPattern,
-                                    private var _input: String,
-                                    private var regionStart: Int,
-                                    private var regionEnd: Int)
+final class RegexMatcher private[regex] (private var _pattern: Pattern,
+                                         private var _input: String,
+                                         private var regionStart: Int,
+                                         private var regionEnd: Int)
     extends AnyRef
     with MatchResult {
 
-  def pattern(): PosixPattern = _pattern
+  def pattern(): Pattern = _pattern
 
   private var inputstr = _input.subSequence(regionStart, regionEnd).toString
 
   def matches(): Boolean =
-    PosixPattern.execute(_pattern, _input) == 0
+    Pattern.execute(_pattern, _input) == 0
 
   def groupCount(): Int = ???
